refractor: timing format is changed to am pm
desc: The timing format is changed from duration to am pm format so that there is no confusion between the time and test finish functionality is added everythinh other is working fine and the flow is established for the project  the timing has little issue of if the test is switching the day. also the time format is also considered brownie point.